add error handling to lib implementation for cli
also fixes some pre-existing bugs with handling unknown
exceptions.
